fix: fixed for PSR-0 loaders.
