+++ tests/namespacetest/polymorphism/Animal.php
@@ -1,11 +1,6 @@
 <?php
 
 
-namespace namespacetest\polimorphism;
-
-
-use InvalidArgumentException;
-
 abstract class Animal
 {
     /** @var string */
@@ -14,9 +9,9 @@ abstract class Animal
     /** @var string */
     public $kind;
 
-    abstract function hasHair(): bool;
+    abstract function hasHair();
 
-    abstract function hasScales(): bool;
+    abstract function hasScales();
 
     public static function jsonMapper($class, $json)
     {
@@ -34,45 +29,7 @@ abstract class Animal
     }
 }
 
-class Cat extends Animal
-{
-    /** @var int */
-    public $legsNumber;
 
-    function hasHair(): bool
-    {
-        return true;
-    }
 
-    function hasScales(): bool
-    {
-        return false;
-    }
-}
 
-class Fish extends Animal
-{
 
-    /** @var bool */
-    public $livesInSaltwater;
-
-    function hasHair(): bool
-    {
-        return false;
-    }
-
-    function hasScales(): bool
-    {
-        return true;
-    }
-}
-
-class Zoo
-{
-    /** @var string */
-    public $name;
-
-    /** @var Animal[] */
-    public $animals;
-
-}
