+++ tests/namespacetest/polymorphism/Cat.php
@@ -1,0 +1,21 @@
+<?php
+
+
+
+
+
+class Cat extends Animal
+{
+    /** @var int */
+    public $legsNumber;
+
+    function hasHair()
+    {
+        return true;
+    }
+
+    function hasScales()
+    {
+        return false;
+    }
+}
