+++ tests/namespacetest/polymorphism/Fish.php
@@ -1,0 +1,21 @@
+<?php
+
+
+
+
+class Fish extends Animal
+{
+
+    /** @var bool */
+    public $livesInSaltwater;
+
+    function hasHair()
+    {
+        return false;
+    }
+
+    function hasScales()
+    {
+        return true;
+    }
+}
