+++ tests/namespacetest/polymorphism/Zoo.php
@@ -1,0 +1,14 @@
+<?php
+
+
+
+
+class Zoo
+{
+    /** @var string */
+    public $name;
+
+    /** @var Animal[] */
+    public $animals;
+
+}
